@@ -1,14 +1,26 @@
 module github.com/ARGOeu/ams-push-server
 
-go 1.14
+go 1.21
+
 
 require (
-	github.com/gogo/protobuf v1.3.2 // indirect
-	github.com/golang/protobuf v1.5.2
-	github.com/grpc-ecosystem/go-grpc-middleware v1.0.0
-	github.com/pkg/errors v0.8.0
-	github.com/sirupsen/logrus v1.2.0
-	github.com/stretchr/testify v1.2.2
-	google.golang.org/genproto v0.0.0-20181109154231-b5d43981345b // indirect
-	google.golang.org/grpc v1.16.0
+	github.com/golang/protobuf v1.5.4
+	github.com/grpc-ecosystem/go-grpc-middleware v1.4.0
+	github.com/pkg/errors v0.9.1
+	github.com/sirupsen/logrus v1.9.3
+	github.com/stretchr/testify v1.8.4
+	google.golang.org/grpc v1.65.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/kr/pretty v0.3.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/net v0.27.0 // indirect
+	golang.org/x/sys v0.22.0 // indirect
+	golang.org/x/text v0.16.0 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20240701130421-f6361c86f094 // indirect
+	google.golang.org/protobuf v1.34.2 // indirect
+	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
