@@ -7,7 +7,7 @@ import (
 	ams "github.com/ARGOeu/ams-push-server/pkg/ams/v1"
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/suite"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"testing"
 )
@@ -196,6 +196,6 @@ func (suite *AmsHttpConsumerTestSuite) TestToCancelableError() {
 }
 
 func TestAmsHttpConsumerTestSuite(t *testing.T) {
-	logrus.SetOutput(ioutil.Discard)
+	logrus.SetOutput(io.Discard)
 	suite.Run(t, new(AmsHttpConsumerTestSuite))
 }
