@@ -4,7 +4,7 @@ import (
 	"context"
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/suite"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"testing"
 )
@@ -86,6 +86,6 @@ func (suite *SubscriptionTestSuite) TestGetSubscription() {
 }
 
 func TestSubscriptionTestSuite(t *testing.T) {
-	logrus.SetOutput(ioutil.Discard)
+	logrus.SetOutput(io.Discard)
 	suite.Run(t, new(SubscriptionTestSuite))
 }
