@@ -5,7 +5,7 @@ import (
 	v1 "github.com/ARGOeu/ams-push-server/pkg/ams/v1"
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/suite"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"testing"
 )
@@ -77,6 +77,6 @@ func (suite *MattermostSenderTestSuite) TestDestination() {
 }
 
 func TestMattermostSenderTestSuite(t *testing.T) {
-	logrus.SetOutput(ioutil.Discard)
+	logrus.SetOutput(io.Discard)
 	suite.Run(t, new(MattermostSenderTestSuite))
 }
